prod-devel-rcar: provide device tree names into DomD
Prior to this patch DomD kernel tried guest required device tree by
looking at SOC and BOARD variables, but this produced multiple DTSes,
as it is not possible to distinguish between, say, 4GB or 8GB
variants.

With this patch, moulin will provide exact list of required DTS files.

As a byproduct, we need to define Xen DTS name as well.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>